Scenario test with anoncreds wallet upgrade and restart
